more type conversion
svn-id: r2270

+++ gemrb/plugins/Core/ImageMgr.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/ImageMgr.h,v 1.7 2004/04/17 11:28:10 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/ImageMgr.h,v 1.8 2004/08/02 22:14:22 guidoj Exp $
  *
  */
 
@@ -48,7 +48,7 @@ public:
 	virtual void GetPalette(int index, int colors, Color* pal) = 0;
 	/** Gets a Pixel from the Image */
 	virtual Color GetPixel(unsigned int x, unsigned int y) = 0;
-	virtual unsigned long GetPixelIndex(unsigned int x, unsigned int y) = 0;
+	virtual unsigned int GetPixelIndex(unsigned int x, unsigned int y) = 0;
 };
 
 #endif
